Extract inline tests + split workflow/register.rs into focused modules (#1425)
Mechanical modularity pass — no behavior changes.

Inline test extractions to sibling files:
- crates/harn-vm/src/orchestration/{records,crystallize}.rs
- crates/harn-cli/src/commands/orchestrator/harness.rs
- crates/harn-cli/src/commands/mcp/serve.rs

workflow/register.rs split (2273 → 200 lines):
- inspect.rs: workflow_graph/validate/clone/insert_node/replace_node/etc.
- hooks.rs: register_tool_hook + register_persona_hook + step hook builtins
- compact.rs: select_artifacts_adaptive/estimate_tokens/microcompact/transcript_auto_compact
- state.rs: workflow run state machine + stage execution scope (1282 lines)
- host.rs: __host_workflow_* registration-layer builtins
